requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
requests==2.32.3
httpx==0.26.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2

# OpenAPI/Schema validation
pyyaml==6.0.2
jsonschema==4.21.1
openapi-spec-validator==0.7.1

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.23.3
